requirements.txt
string
os
wordcloud
matplotlib.pyplot
random
numpy
nltk